Dagger setup

+++ domain/src/main/java/net/chmielowski/shoppinglist/types.kt
@@ -1,3 +1,3 @@
 package net.chmielowski.shoppinglist
 
-typealias Id = Int
+typealias Id = Long
